Fixed ANE compilation to allow building on Windows machines (Fixes issue #23)
Changed the share dialog parameters to only accept a single friend ID as a String rather than an array of friend IDs, which is not actually supported. (Fixes issue #48)

Added userHasIntegratedFacebookAccount() method to detect if a user has a Facebook account connected in iOS settings. (Issue #26)

Added support for logging Facebook app events and purchases (Issue #13). See the documentation for more info.
New methods include:
- logEvent
- logPurchase
- setAppVersion (sets an app version string to be associated with logged events)
- setAppEventFlushBehavioud
- flushAppEvents

Added appActivate() event to be in line with new Facebook SDK, instead of the deprecated publishInstall method.